+++ src/main/java/com/ecommerce/springbootecommerce/api/seller/ProductAPI.java
@@ -1,13 +1,20 @@
 package com.ecommerce.springbootecommerce.api.seller;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.propertyeditors.StringTrimmerEditor;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -22,7 +29,6 @@ import org.springframework.web.servlet.view.RedirectView;
 import com.ecommerce.springbootecommerce.dto.ProductDTO;
 import com.ecommerce.springbootecommerce.service.IProductService;
 
-
 @RestController
 @RequestMapping("/api/seller")
 public class ProductAPI {
@@ -32,23 +38,40 @@ public class ProductAPI {
 
     public static String uploadDirectory = System.getProperty("user.dir") + "/src/main/resources/imagedata/seller";
     
+    @InitBinder
+    public void initBinder(WebDataBinder dataBinder) {
+        StringTrimmerEditor stringTrimmerEditor = new StringTrimmerEditor(true);
+        dataBinder.registerCustomEditor(String.class, stringTrimmerEditor);
+    }
+    
     @PostMapping(value="/product")
     public RedirectView createProduct(
-        @ModelAttribute(name = "product") ProductDTO product,
-        @RequestParam(value = "imageField") MultipartFile file
+        @Valid @ModelAttribute(name = "product") ProductDTO product,
+        @RequestParam(value = "imageField") MultipartFile file,
+        BindingResult bindingResult
     ) throws IOException{
+        
+        if (bindingResult.hasErrors()) {
+            return new RedirectView("seller/product/newProduct");
+        }
+        
         byte[] imageBytes = file.getBytes();
         
         String author = SecurityContextHolder.getContext().getAuthentication().getName();
         
         String fileName= file.getOriginalFilename();
-        Path fileNameAndPath = Paths.get(uploadDirectory, author + "_" + fileName);
+        Path storePath = Paths.get(uploadDirectory, author);
+        Path fileNameAndPath = Paths.get(uploadDirectory + "/" + author, fileName);
         
-        try {
+        File directoryStorePath = new File(String.valueOf(storePath));
+        File directoryFileNamePath = new File(String.valueOf(fileNameAndPath));
+                
+        if(!directoryStorePath.exists()) {    
+            Files.createDirectories(storePath);
+        }
+        
+        if(!directoryFileNamePath.exists()) {    
             Files.write(fileNameAndPath, imageBytes);
-
-        } catch (Exception e) {
-            e.printStackTrace();            
         }
 
         product.setImage(imageBytes);
